1.4.53 14Mar2019 AQU=077a Last code before I clean up much of the tools to try to catch the xTIMEcomposer 14.3.3 tool problem. This code runs
14Mar2019 AQU=077  A replacement for AQU=065?, see below. I DECIDE THAT
IT IS OUT OF MY CONTROL TO FIX THIS AS IT MUST BE A PROBLEM WITH
xTIMEcomposer 14.3.3. Therefore I will clean up the code that tries try
to catch AQU=077 and try to see if TRANS_ASYNCH_WRAPPED may (after all)
be a method to isolate the AQU=065 problem. And let the aquarium run
with that for a while.

14Mar2019 AQU=065? It is not certain that the original AQU=065 stop and
the fact that the AQU=065letter entries below are the same problem! The
latter is about the fact that (on the startKIT?, some specific par
placement?, some sequence?) do not seem to be able to exit from some
early i_radio.nn interface call. See llts of log and scope pictures.
However, the original AQU=065 does survive for days before a stop is
seen. In both cases the SW starts to beep every 10 seconds in
Port_Pins_Heat_Light_Task.